Migrate to existing typings in `jimp` package
