+++ packages/plugins/src/index.d.ts
@@ -1,8 +1,1 @@
-export interface Plugin<ClassT> {
-  class?: ClassT;
-  constants: {
-    [key: string]: string
-  }
-}
-
 export default function(jimpEvChange: any): Jimp;
