+++ packages/type-png/src/index.d.ts
@@ -1,16 +1,14 @@
-import Jimp = require('jimp');
-
-import { ITypePluginReturn, Image, jimpCB } from '../../types/src';
+import Jimp, {ImageCallback, ITypePluginReturn, Image} from 'jimp';
 
 interface PNGImage extends Image {
     _deflateLevel: number,
     _deflateStrategy: number,
     _filterType: number,
     _colorType: number,
-    deflateLevel(l: number, cb?: jimpCB): Jimp;
-    deflateStrategy(s: number, cb?: jimpCB): Jimp;
-    filterType(f: number, cb?: jimpCB): Jimp;
-    colorType(s: number, cb?: jimpCB): Jimp;
+    deflateLevel(l: number, cb?: ImageCallback): Jimp;
+    deflateStrategy(s: number, cb?: ImageCallback): Jimp;
+    filterType(f: number, cb?: ImageCallback): Jimp;
+    colorType(s: number, cb?: ImageCallback): Jimp;
 }
 
 interface TypePngRet extends ITypePluginReturn<PNGImage> {
